+++ agents/rag.py
@@ -2,23 +2,29 @@
 
 import logging
 from operator import itemgetter
-from typing import Any, List
+from typing import Any, List, Optional
 
+from langchain.schema import Document
 from langchain.text_splitter import RecursiveCharacterTextSplitter
 from langchain_community.vectorstores import FAISS
 from langchain_core.output_parsers import StrOutputParser
 from langchain_core.prompts import ChatPromptTemplate
 from langchain_openai import ChatOpenAI, OpenAIEmbeddings
 
+from config.config import FileUploadConfig
 from services.document_service import load_documents
+from services.url_service import WebScraper
 
 
-def setup_rag_chain(file_paths: List[str], llm: ChatOpenAI) -> Any:
-    """Set up a Retrieval-Augmented Generation (RAG) chain using loaded documents and save the resulting vector index."""
-    logging.info(f"Setting up RAG chain for files: {file_paths}")
+def setup_rag_chain(files_path_list: List[str], url: str, llm: ChatOpenAI) -> Any:
+    """
+    Set up a Retrieval-Augmented Generation (RAG) chain using either loaded
+    documents or documents extrcted from url and save the resulting vector index.
+    """
+    logging.info(f"Setting up RAG chain for: {files_path_list or url}")
 
     # Load documents using the new document_loader module
-    docs = load_documents(file_paths)
+    docs = get_documents(files_path_list, url)
 
     if docs:
         text_splitter = RecursiveCharacterTextSplitter(chunk_size=300, chunk_overlap=0)
@@ -43,3 +49,21 @@ def setup_rag_chain(file_paths: List[str], llm: ChatOpenAI) -> Any:
         return rag_chain
     else:
         raise ValueError("No documents were loaded, RAG chain setup cannot proceed.")
+
+
+def get_documents(
+    files_uploaded: Optional[List[str]] = None, url: Optional[str] = None
+) -> List[Document]:
+    """Load documents from either files uploads or a URL."""
+    documents = []
+
+    if files_uploaded:
+        logging.info(f"Loading documents from file uploads: {files_uploaded}")
+        documents.extend(load_documents(files_uploaded))
+
+    if url:
+        logging.info(f"Loading documents from URL: {url}")
+        web_scraper = WebScraper()
+        documents.extend(web_scraper.scrape_website(url))
+
+    return documents
